Share close_scheduled_coro / record_scheduled_coros across mDNS suites
Replace four copies of the `create_background_task` side-effect (in
test_dns_cache, test_mdns_mac_address, test_mdns_version,
test_state_fanout_duplicate_names) with the two shared helpers in
tests/conftest.py. Net -36 lines.
